Change UI for Login, Signup, Home, Restaurant, FoodDetails. Make CollapsingToolbar.

+++ composeApp/src/commonMain/kotlin/com/foodsaver/app/presentation/featureHome/components/HomeTopBar.kt
@@ -44,7 +44,7 @@ fun HomeTopBarPreview() {
 @Composable
 fun HomeTopBar(
     deliverTo: String,
-    cartItemValue: Int?,
+    cartItemValue: Long?,
     onCartClick: () -> Unit,
     onMenuClick: () -> Unit,
     modifier: Modifier = Modifier,
